question 2 to 5 with new list class
list in class and Root, all answers

+++ WebAPIClient/Coord.cs
@@ -2,6 +2,7 @@ namespace WebAPIClient;
 
 public class Coord
 {
-    public int lon { get; set; }
-    public int lat { get; set; }
+    public float lon { get; set; }
+    public float lat { get; set; }
 }
+
